added description of services and middleware for error handling

+++ controllers/noteController.js
@@ -1,31 +1,58 @@
 const NoteService = require('../services/noteService');
+const ApiError = require('../handlers/errors');
 
 class NoteController {
   async create(req, res, next) {
-    const note = await NoteService.create(req.body);
-    return res.json(note);
+    try {
+      const note = await NoteService.create(req.body);
+      return res.json(note);
+    } catch (err) {
+      return next(ApiError.internalServer(err));
+    }
   }
 
   async getAll(req, res, next) {
-    const notes = await NoteService.getAll();
-    return res.json(notes);
+    try {
+      const notes = await NoteService.getAll();
+      return res.json(notes);
+    } catch (err) {
+      return next(ApiError.internalServer(err));
+    }
   }
 
   async getOne(req, res, next) {
-    const { id } = req.params;
-    const note = await NoteService.getOne(id);
-    return res.json(note);
+    try {
+      const { id } = req.params;
+      if (!id) {
+        return next(ApiError.badRequest('не передан параметр id'));
+      }
+      const note = await NoteService.getOne(id);
+      return res.json(note);
+    } catch (err) {
+      return next(ApiError.internalServer(err));
+    }
   }
 
   async update(req, res, next) {
-    const updatedNote = await NoteService.update(req.body);
-    return res.json(updatedNote);
+    try {
+      const updatedNote = await NoteService.update(req.body);
+      return res.json(updatedNote);
+    } catch (err) {
+      return next(ApiError.internalServer(err));
+    }
   }
 
   async delete(req, res) {
-    const { id } = req.params;
-    const note = await NoteService.delete(id);
-    return res.json(note);
+    try {
+      const { id } = req.params;
+      if (!id) {
+        return next(ApiError.badRequest('не передан параметр id'));
+      }
+      const note = await NoteService.delete(id);
+      return res.json(note);
+    } catch (err) {
+      return next(ApiError.internalServer(err));
+    }
   }
 }
